temp

+++ src/Point/Point.hpp
@@ -3,14 +3,14 @@
 
 class Point {
     private:
-        long value;
+        long long value;
     public:
         Point(); // default ctor, value = 64
-        Point(long); // ctor
+        Point(int); // ctor
 
         // getter / setter (don't set value directly, use for resetting game point only)
         // use Add instead for adding value
-        long getValue() const;
+        long long getValue() const;
         void setValue(int);
 
         /* value modification from player skills */
@@ -19,7 +19,7 @@ class Point {
         void Quadruple();
         void Quarter();
 
-        void Add(long);
+        void Add(int);
 };
 
 #endif
